+++ src/axolotl/cli/merge_lora.py
@@ -8,9 +8,10 @@ import fire
 import transformers
 from dotenv import load_dotenv
 
+from axolotl.cli.args import TrainerCliArgs
 from axolotl.cli.art import print_axolotl_text_art
 from axolotl.cli.config import load_cfg
-from axolotl.common.cli import TrainerCliArgs, load_model_and_tokenizer
+from axolotl.cli.utils import load_model_and_tokenizer
 from axolotl.utils.dict import DictDefault
 
 LOG = logging.getLogger(__name__)
@@ -31,10 +32,7 @@ def do_merge_lora(*, cfg: DictDefault) -> None:
 
     LOG.info("Running merge of LoRA with base model...")
     model = model.merge_and_unload(progressbar=True)
-    try:
-        model.to(dtype=cfg.torch_dtype)
-    except RuntimeError:
-        pass
+    model.to(dtype=cfg.torch_dtype)
     model.generation_config.do_sample = True
 
     if cfg.local_rank == 0:
